fix: keyboard-operable listening-language picker, split into own component files
